% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsim.R
\name{adsim}
\alias{adsim}
\title{adsim}
\usage{
adsim(ngens, k, l, admix, m1, m2, ne, p1, p2, nsims)
}
\arguments{
\item{ngens}{Number of generations to simulate.}

\item{k}{population growth parameter.}

\item{l}{number of generations prior to population growth.}

\item{admix}{Proportion of ancestry contributed by the first parent
population.}

\item{m1}{Migration rate from the first parent population.}

\item{m2}{Migration rate from the second parent population.}

\item{ne}{effective population size of generation 0.}

\item{p1}{Initial allele frequency in the first parent population.}

\item{p2}{Initial allele frequency in the second parent population.}

\item{nsims}{Number of simulations to perform.}
}
\value{
A data frame containing three columns:
\itemize{
\item Sim: a numeric index containing the simulation run
\item p0: a numeric containing the starting allele following the first generation
of admixture.
\item pF: a numeric containing the final allele frequency of each simulation.
}
}
\description{
This function simulates evolution of an allele through genetic drift in an
admixed population. The user must specify the starting allele frequency in
both parent populations, the number of generations to simulate, the starting
effective population size, the admixture proportion, the population growth
parameters, the migration rate from both parent populations, and the number
of simulations to perform. The function returns a data frame containing the
final allele frequency of each simulation.The function accepts lists as input
as well.
}
