% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsimR-package.R
\docType{package}
\name{adsimR-package}
\alias{adsimR}
\alias{adsimR-package}
\title{adsimR: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Evie Carter \email{ejc199@pgr.aru.ac.uk} [copyright holder]

Authors:
\itemize{
  \item Evie Carter \email{ejc199@pgr.aru.ac.uk} [copyright holder]
  \item Jael Vakil Kumar
  \item Petra Dozie
  \item Jason Hodgson
}

}
\keyword{internal}
