% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_values.R
\name{p_values}
\alias{p_values}
\title{p_value}
\usage{
p_values(df, observed_freq)
}
\arguments{
\item{df}{A data frame generated by the adsim() function}

\item{observed_freq}{Observed allele frequency}
}
\value{
a data frame showing observed allele frequencies, expected allele
frequencies, z-scores, and p-values for each combination of p1, p2, and admixture
proportion
}
\description{
This function calculates p-values for observed vs expected allele frequencies
}
\examples{
test <- adsim(ngens = 10, k = 2, l = 5, admix = c(0.2, 0.5),
m1 = 0.01, m2 = 0.01, ne = 100, p1 = 0.5,
p2 = 0.9, nsims = 10)

obs_freq <- 0.9

p_values(test, obs_freq)
}
