% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_simulations.R
\name{plot_simulations}
\alias{plot_simulations}
\title{plot_simulations}
\usage{
plot_simulations(adsim_df)
}
\arguments{
\item{adsim_df}{A data frame containing three columns:
\itemize{
\item Sim: a numeric index containing the simulation run
\item p0: a numeric containing the starting allele following the first
generation of admixture.
\item pF: a numeric containing the final allele frequency of each simulation.
}}
}
\value{
a plot containing two panes:
\itemize{
\item a plot containing the summary of allele frequencies of the simulation,
including quartiles at 75\%, 90\%, and 95\%.
\item a table containing the parameters of each simulation.
}
}
\description{
This function creates an autoplot of adsim output
}
\examples{
test <- adsim(ngens = 10, k = 2, l = 5, admix = c(0.2, 0.5),
m1 = 0.01, m2 = 0.01, ne = 100, p1 = 0.5,
p2 = 0.9, nsims = 10)

plot_simulations(test)
}
