% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_overview.R
\name{sim_overview}
\alias{sim_overview}
\title{sim_overview}
\usage{
sim_overview(adsim_df)
}
\arguments{
\item{adsim_df}{A data frame containing three columns:
\itemize{
\item Sim: a numeric index containing the simulation run
\item p0: a numeric containing the starting allele following the first
generation of admixture.
\item pF: a numeric containing the final allele frequency of each simulation.
}}
}
\value{
a data frame summarizing the mean, standard deviation, and percentiles
}
\description{
This function creates an overview summary of adsim output
}
\examples{
test <- adsim(ngens = 10, k = 2, l = 5, admix = c(0.2, 0.5),
m1 = 0.01, m2 = 0.01, ne = 100, p1 = 0.5,
p2 = 0.9, nsims = 10)

sim_overview(test)
}
