% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_value.R
\name{z_value}
\alias{z_value}
\title{z_value}
\usage{
z_value(df)
}
\arguments{
\item{df}{A data frame generated by the adsim() function}
}
\value{
a data frame showing observed allele frequencies, expected allele
frequencies, and z-scores for each combination of p1, p2, and admixture
proportion
}
\description{
This function calculate z-scores for observed vs expected allele frequencies
}
\examples{
test <- adsim(ngens = 10, k = 2, l = 5, admix = c(0.2, 0.5),
m1 = 0.01, m2 = 0.01, ne = 100, p1 = 0.5,
p2 = 0.9, nsims = 10)
z_value(test)
}
