% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_selection_coefficient.R
\name{compute_selection_coefficients}
\alias{compute_selection_coefficients}
\title{Compute selection coefficients from allele frequency scenarios}
\usage{
compute_selection_coefficients(infile, outfile)
}
\arguments{
\item{infile}{Path to input file (columns: g, q, qx).}

\item{outfile}{Path to save output table (with added column s).}
}
\value{
Data frame with selection coefficients.
}
\description{
Compute selection coefficients from allele frequency scenarios
}
\examples{
result <- compute_selection_coefficients(infile = system.file("perl_base_code/Selection_scenarios.txt", package = "adsimR"),
outfile = paste0(tempdir(), "/output.txt"))
}
