% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_selection_coefficient.R
\name{compute_selection_coefficients}
\alias{compute_selection_coefficients}
\title{Compute selection coefficients from allele frequency scenarios}
\usage{
compute_selection_coefficients(infile, outfile, recessive = FALSE)
}
\arguments{
\item{infile}{Either a data frame or path to input file, containing the
following columns:
\itemize{
\item g generations
\item q: initial allele frequency
\item qx: final allele frequency
}}

\item{outfile}{Path to save output table. This will include added column s
giving the selection coefficient.}

\item{recessive}{Logical, if TRUE assumes recessive model, if FALSE assumes
dominant model. Default is FALSE.}
}
\value{
A data frame with selection coefficients, containing the columns:
\itemize{
\item g: generations
\item q: initial allele frequency
\item qx: final allele frequency
\item s: selection coefficient
}
}
\description{
This function takes a data frame or a path to an input file containing allele
frequency scenarios and computes the selection coefficients based on the
provided data.
}
\examples{
result <-
      compute_selection_coefficients(infile = system.file("extdata",
                                                          "Selection_scenarios.txt",
                                                           package = "adsimR"),
                                   outfile = paste0(tempdir(), "/output.txt"))
}
