% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_selection_coefficient.R
\name{compute_selection_coefficients}
\alias{compute_selection_coefficients}
\title{Compute selection coefficients from allele frequency scenarios}
\usage{
compute_selection_coefficients(infile, outfile, recessive = FALSE)
}
\arguments{
\item{infile}{Path to input file (columns: g, q, qx).}

\item{outfile}{Path to save output table (with added column s).}

\item{recessive}{Logical, if TRUE assumes recessive model, if FALSE assumes
dominant model. Default is FALSE.}
}
\value{
Data frame with selection coefficients.
}
\description{
Compute selection coefficients from allele frequency scenarios
}
\examples{
result <- compute_selection_coefficients(infile = system.file("extdata","Selection_scenarios.txt", package = "adsimR"),
outfile = paste0(tempdir(), "/output.txt"))
}
