% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_values.R
\name{p_values}
\alias{p_values}
\title{p_value}
\usage{
p_values(adsim_df, observed_freq)
}
\arguments{
\item{adsim_df}{A data frame containing the output of an \code{adsim} simulation,
with the following columns:
\itemize{
\item Sim: a numeric index containing the simulation run
\item p0: a numeric containing the starting allele frequency following the first
generation of admixture.
\item pF: a numeric containing the final allele frequency of each simulation.
\item ngens: Number of generations simualted.
\item k: population growth parameter
\item l: number of generations prior to population growth.
\item admix: Proportion of ancestry contributed by the first parent
population.
\item ne: effective population size of generation 0.
\item p1: Initial allele frequency in the first parent population.
\item p2: Initial allele frequency in the second parent population.
\item m1: Migration rate from the first parent population.
\item m2: Migration rate from the second parent population.
}}

\item{observed_freq}{Observed allele frequency}
}
\value{
A data frame returning a p-value and a z-score for the observed
allele frequency compared to the expected allele frequency derived from the
simulated dataset. The output contains one row for each unique set of
simulated parameters, with the following columns:
\itemize{
\item p1: Initial allele frequency in the first parent population.
\item p2: Initial allele frequency in the second parent population.
\item admix: Proportion of ancestry contributed by the first parent
population.
\item exp_freq: Expected allele frequency based on the admixture proportion and
the initial allele frequencies of the two parent populations.
\item z: z-score calculated as (observed_freq - exp_freq) / sd_pF,
where exp_freq is based on the admixture proportion and the initial allele
frequencies of the two populations, and sd_pF is the standard deviation of
the final allele frequencies across all simulations.
\item p: p-value calculated as 2*(1 - pnorm(abs(z))), where z is the z-score
described in the documentation for \code{\link[=z_value]{z_value()}}.
}
}
\description{
This function calculates p-values for observed vs expected allele
frequencies.
}
\examples{
test <- adsim(ngens = 10, k = 2, l = 5, admix = c(0.2, 0.5),
m1 = 0.01, m2 = 0.01, ne = 100, p1 = 0.5,
p2 = 0.01, nsims = 10)

obs_freq <- 0.9

p_values(test, obs_freq)
}
